fix(agent): macOS build — statvfs field widths and pipe2 fallback
- statvfs.f_blocks/f_bavail are u32 on macOS but u64 on Linux; normalize with
  'as u64' before the multiplication.
- pipe2(2) does not exist on macOS; use pipe(2) on non-Linux unix.

+++ src/agent-rs/libra-platform/src/linux.rs
@@ -176,8 +176,8 @@ impl IPlatformExecutor for LinuxExecutor {
                     ) == 0
                     {
                         let bsize = stat.f_frsize as u64;
-                        total = stat.f_blocks.saturating_mul(bsize);
-                        free = stat.f_bavail.saturating_mul(bsize);
+                        total = (stat.f_blocks as u64).saturating_mul(bsize);
+                        free = (stat.f_bavail as u64).saturating_mul(bsize);
                     }
                 }
                 if total == 0 {
